src/requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
Babel==2.12.1
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==15.1.0
cron-descriptor==1.4.0
cryptography==41.0.3
daphne==4.0.0
Django==4.2.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-money==3.2.0
django-timezone-field==5.1
django-webpush==0.3.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.2
frozenlist==1.4.0
future==0.18.3
h11==0.14.0
http-ece==1.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
kombu==5.3.1
msgpack==1.0.5
multidict==6.0.4
nmaptocsv==1.6
outcome==1.2.0
prompt-toolkit==3.0.39
psycopg2==2.9.7
psycopg2-binary
py-moneyed==3.0
py-vapid==1.9.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.2.0
PySocks==1.7.1
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
pywebpush==1.9.4
redis==4.6.0
requests==2.31.0
selenium==4.11.2
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.22.2
trio-websocket==0.10.3
twilio==8.5.0
Twisted==22.10.0
#twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
wsproto==1.2.0
yarl==1.9.2
zope.interface==6.0
